Fix two issues in the thread safety analysis.
Fix two issues in the thread safety analysis. First, it uses DECL_CONTEXT
of a function decl node to determine whether the decl is a class member
instead of checking if the function decl type is METHOD_TYPE, which
doesn't work for static members. Second, when comparing locks, we need
to build and use fully-qualified names if the locks are class members.
Also added new test cases.

From-SVN: r138754