Reduce the initial size of the libcpp identifier lookaside table to 2^9 slots from 2^15 slots.
Reduce the initial size of the libcpp identifier lookaside table to
2^9 slots from 2^15 slots.  This change exercises the table growth.

When capturing an identifier list, reduce the size of the lookaside
table to the origonally requested size, as long as that saves more
than half the space.  Otherwise, it is not worth doing.

Make documentation and code clarifications.

From-SVN: r167669