An earlier change reduced the number of bits in cpp_hashnode.directive_index from 7 to 5...
An earlier change reduced the number of bits in cpp_hashnode.directive_index
from 7 to 5, as that was sufficient for indexing the directives.  Tom Tromey
asked for a static check on the size.  This patch adds that check.

Unfortunately, five bits are not sufficient for the alternate use of
cpp_hashnode.directive_index as a named operator index.  So, I have reverted
the number of bits from five back to seven.  As a result, we now have 34 bits
in small fields, and the size of cpp_hashnode will increase from two to three
words on 32-bit systems.  The size on 64-bit systems remains unchanged because
these bits go into an alignment gap.

Index: libcpp/ChangeLog.pph

2011-04-14  Lawrence Crowl  <crowl@google.com>

	* include/cpplib.h (cpp_hashnode):  Use a macro for the number of bits
	in the directive_index bitfield.  Change the number of bits back to 7.
	* directives.c (DIRECTIVE_TABLE):  Add a gcc-build-time check that the
	index of the directive table fits within the number of bits above.
	* init.c (operator_array):  Separate the named operator table into a
	macro.  Use it in operator_array.  Test that the values fit within the
	number of bits above.

From-SVN: r172514