The purpose of this patch is to make PPH testing clean.
Some tests have assembly comparison failures between regular compiles and PPH
compiles.  When this miscompare is exected, we now mark the test with a
comment "pph asm xdiff".  The driver lib/dg-pph.exp looks for this comment to
report either FAIL or XFAIL.

Test p1mean.cc enters infinite recursion.  Instead of disabling the test with
a syntax error, we now run it under a dg-timeout.

Test c1return-5.h was trying to grep assembly, but since we produce pre-parsed
headers, not assembly, the test was in error.  We have moved the assembly
check to c1return-5.cc where it belongs.

All PPH tests expected to have PPH compilation failures are marked with
// { dg-xfail-if "comment" { "*-*-*" } { "-fpph-map=pph.map" } }
The regular compiles do not have the -fpph-map option, and hence are not
expected to fail.  One test has a comment INFINITE; one test has a comment
BOGUS; and the rest have a comment ICE.

The driver lib/dg-pph.exp test control logic now stops a test when any
compilation fails to produce assembly output.  If the test file contains the
above directive, that compilation failure is expected and reported as such.
Otherwise, it is reported as an unexpected failure.

PPH is wrongly producing ICEs and errors.  We mark applicable messages with
dg-bogus.  Because the line/column information is often wrong, we mark the
with a line of "0", which avoids line comparisons.  We also prune output as
needed.

All tests are now either PASS or XFAIL, with the exception of unavoidable
bogus XPASS messages for "test for bogus messages" and "test for excess
errors".  These are unavoidable because they cannot be conditionalized based
on compiler options, and they do not appear in regular compiles.

If you run the output of

    make check-c++ RUNTESTFLAGS=pph.exp

through

    sed -e '
        /^XPASS: .*test for bogus messages/     d
        /^XPASS: .*test for excess errors/      d

        /^#/            p
        /^ERROR: /      p
        /^XFAIL: /      p
        /^XPASS: /      p
        /^FAIL: /       p

        d
    '

you should get no individual tests listed.


Index: gcc/testsuite/ChangeLog.pph

2011-06-05  Lawrence Crowl  <crowl@google.com>

	* lib/dg-pph.exp (dg-pph-hdr): Add blank lines to logfile.
	Format if statements and calls consistently.
	(dg-pph-neg): Remove extraneous assembly files. Add blank
	lines to logfile. Format if statements and calls consistently.
	(dg-pph-pos): Stop test when a compilation fails. Recognize
	expected failing compilations. Recognize expected assembly
	miscompares. Parenthesize failure notes. Remove extraneous
	assembly files. Add blank lines to logfile. Format if
	statements and calls consistently.
	* g++.dg/pph/c1simple.cc: Add comment "pph asm xdiff"
	indicating expected assembly difference failure. Add line
	spacing to test.
	* g++.dg/pph/c1struct.cc: Likewise.
	* g++.dg/pph/c1variables.cc: Likewise.
	* g++.dg/pph/c2builtin1.cc: Likewise.
	* g++.dg/pph/x1globalref.cc: Likewise.
	* g++.dg/pph/x1hardlookup.cc: Likewise.
	* g++.dg/pph/x1struct1.cc: Likewise.
	* g++.dg/pph/x1struct2.cc: Likewise.
	* g++.dg/pph/x2nontrivinit.cc: Likewise.
	* c120060625-1.h: Add dg-xfail-if for -fpph-map. Add bogus
	message for ICE. Prune output. Add line spacing to text.
	* g++.dg/pph/c120060625-1.cc: Likewise.
	* g++.dg/pph/c1eabi1.h: Likewise.
	* g++.dg/pph/c1eabi1.cc: Likewise.
	* g++.dg/pph/x1autometh.cc: Likewise.
	* g++.dg/pph/x1functions.cc: Likewise.
	* g++.dg/pph/x1special.cc: Likewise.
	* g++.dg/pph/x1template.cc: Likewise.
	* g++.dg/pph/x1tmplclass.cc: Likewise.
	* g++.dg/pph/x1tmplfunc.cc: Likewise.
	* g++.dg/pph/x1typerefs.cc: Likewise.
	* g++.dg/pph/x1variables.cc: Likewise.
	* g++.dg/pph/p1mean.cc: Change syntax error preventing infinite
	recursion to a dg-timeout. Add dg-xfail-if for -fpph-map.
	* g++.dg/pph/c1return-5.h: Remove dg-final as we do not generate
	assembly on compiling headers. Change dg directive location.
	* g++.dg/pph/c1return-5.cc: Add dg-final as we _do_ generate
	assembly on compiling sources. Copy up dg-options from
	c1return-5.h as well.

From-SVN: r174669