This patch modifies the behavior of cp_get_virtual_function_decl in...
This patch modifies the behavior of cp_get_virtual_function_decl in
gcc/cp/class.c so that it returns NULL if the function declaration
cannot be found.  The previous behavior was to fail with a
segmentation fault.  The method-not-found case may occur when
Annotalysis uses the function to look up a method, in cases where the
static type cannot be accurately determined.

From-SVN: r176188