This patch fixes get_canonical_lock_expr so that it works on lock expressions that involve a MEM_REF.
This patch fixes get_canonical_lock_expr so that it works on lock
expressions that involve a MEM_REF.  The canonical form of a lock 
expression must rewrite all MEM_REFs to INDIRECT_REFs to accurately 
compare expressions.  It also changes lock_expr_tab so that it is 
allocated by the garbage collector, and properly marked as a GC root.

2011-07-19  DeLesley Hutchins <delesley@google.com>

        * tree-threadsafe-analyze.c:  Changes lock_expr_tab to be 
        allocated bythe garbage collector, and marked as a GC root.
        * tree-threadsafe-analyze.c (get_canonical_lock_expr):  
        Changed to force rewrite on MEM_REF
        * thread_annot_lock-80.C:  Additional test cases for handling 
        MEM_REF expressions.

From-SVN: r176489