In some translation units the debug location of make_heap is the location of its name...
In some translation units the debug location of make_heap is the
location of its name; in other TUs it's the location of the closing
')'. This causes false positives in gold's ODR checker. Until I can
find why the locations are inconsistent, we can work around the
problem by putting the name and closing ')' on the same line.

	Google ref 5068684.
	* libstdc++-v3/include/bits/stl_heap.h(make_heap): Remove a newline.

From-SVN: r176716