This patch fixes an ICE caused when the ipa-sra optimization deletes function...
This patch fixes an ICE caused when the ipa-sra optimization deletes
function arguments that are referenced from within a thread safety
attribute.  It will attempt to detect this situation and recover
gracefully.  Since a graceful recovery is not always possible, an
optional warning (-Wwarn-thread-optimization) can be turned on to
inform the user that certain attributes have been removed by
optimization.

From-SVN: r180984