My previous patch moved the in_memory marker to the wrong place.
We need to prevent a file from being re-read by an #include from one
of its included children images.  This is common in system headers.

We mark the file as resident in-memory right before we start reading
its includes.  Before, we were marking it when the file was initially
registered.  This was causing us to completely ignore needed
PPH images.

	* pph-core.c (pph_stream_register): Move in-memory marking...
	* pph-in.c (pph_read_file_1): ... here.

From-SVN: r183256