When we read a tree from the cache of an external PPH file and the tree is mutated in the current file...
When we read a tree from the cache of an external PPH file and the
tree is mutated in the current file, we need to re-sign the tree to
reflect the changes.

The problem was, that the signing code was assuming that the tree had
come from the *current* file, so it was trying to sign a cache entry
for the internal cache, which would ICE if the internal cache is
smaller than the external one.

I have not yet been able to produce a small test case for this, but it
fixes several thousand ICEs in our internal code.  I will add a test
case as soon as I get one reduced.

	* pph-in.c (pph_in_tree): Sign EXPR from the cache it was
	taken from.
	* pph-streamer.h (pph_cache_select): Handle PPH_RECORD_START,
	PPH_RECORD_START_MERGE_BODY, PPH_RECORD_START_MERGE_KEY
	and PPH_RECORD_START_MUTATED.

From-SVN: r183311