This patch fixes an ICE in build_int_cst_wide.
This patch fixes an ICE in build_int_cst_wide.  We were not writing
out TYPE_MODE for types, causing RTL generation to try to emit an
offset using VOIDmode.

This is not the only problem in this test case.  We are also failing
to mark something needed, because with PPH enabled, a whole function
is optimized out.  I'm not quite sure why yet.  I will fix this in a
subsequent patch.

2012-01-20   Diego Novillo  <dnovillo@google.com>

cp/ChangeLog.pph
	* pph-in.c (pph_in_machine_mode): New.
	(pph_in_tcc_type): Call it.
	* pph-out.c (pph_out_machine_mode): New.
	(pph_out_tcc_type): Call it.

testsuite/ChangeLog.pph
	* g++.dg/pph/x1namespace-alias2.cc: Remove expected ICE.
	Document assembly difference.

From-SVN: r183364