This patch factors out some code, which was making debugging unnecessarily confusing.
When we read the header of a tree, we first read an LTO tag marker to
identify what kind of tree we are reading.  However, we were doing
this *outside* of pph_in_tree_header, which resulted in code
duplication in the several callers we have for it.

The patch also removes some unnecessary code:

- In pph_out_tree, we will never have an IDENTIFIER_NODE when writing
  a builtin.  Similarly, in pph_in_tree.

- There was some code commented out in
  pph_in_merge_body_namespace_decl that is really not needed anymore.
  We are now treating mutated trees as XREFs, instead of trying to add
  them on the internal cache.

2012-01-24   Diego Novillo  <dnovillo@google.com>

	* pph-in.c (pph_in_tree_header): Add argument FULLY_READ_P.
	Remove argument TAG.  Update all callers.
	Move LTO tag reading code from ...
	(pph_in_tree): ... here.
	(pph_in_merge_body_namespace_decl): Remove unused code.
	* pph-out.c (pph_out_tree): Remove check for IDENTIFIER_NODE.

From-SVN: r183495