A small fixlet to deal with files that have errors in them.
A small fixlet to deal with files that have errors in them.  When we
are about to generate PPH output, we check if there have been errors
emitted for the file.  If so, we disable PPH generation.

Testing this, I found that we were not clearing pph_out_stream
when disabling PPH.  Fixed.

2012-01-26   Diego Novillo  <dnovillo@google.com>

	* pph-core.c (pph_stream_close_1): Add a newline to fprintf call.
	(pph_finish): If there were errors during compilation,
	disable PPH generation.
	* pph-out.c (pph_disable_output): Also set pph_out_stream to
	NULL.

From-SVN: r183578