When generating PPH images...
When generating PPH images, circular #includes cannot be handled using
the usual double-guard inclusion in the pre-processor.  This happens
because we generate PPH images in separate compilations.  So, even
with properly double-guarded headers, each header will include the
other in its image:

1.h

2.h

When compiling 1.h, we will try open 2.pph, which in turn includes
1.h, which is being generated.  To avoid this situation, we simply
force the inclusion of 1.h to proceed as a text inclusion.

When adding a test for this, I ran into trouble with the testing
harness.  The test would fail depending on which PPH images were
already present in the build directory.  So, I modified pph.exp to
remove all the PPH images before starting the main test loop.  I think
this will not work in a remote test setting, but I will deal with that
later.

2012-01-27   Diego Novillo  <dnovillo@google.com>

cp/ChangeLog.pph
	* pph-core.c (pph_stream_open): If STREAM exists and its mode
	is different than the mode in which it was originally opened,
	return NULL.

testsuite/ChangeLog.pph
	* g++.dg/pph/pph.exp: Remove PPH images before running the tests.
	* g++.dg/pph/x1circular.h: New.
	* g++.dg/pph/x2circular.h: New.
	* g++.dg/pph/x3circular.cc: New.

From-SVN: r183640