This patch adds support for disabling PPH reads.
This patch adds support for disabling PPH reads.  It uses the new
support to disable PPH reading when we disable PPH writing.  This is
to avoid confusion in the PPH registration code that will have some
stale content after the output file has been removed.

Lawrence suggests that we should just fail hard when we cannot
generate a PPH image.  We are currently only emitting a warning.  I
think I will just rip out this support in the near future.  For now,
it was easier to just disable the reader.  For now, I'm leaving it in
just in case we have another need for this.

2012-01-27   Diego Novillo  <dnovillo@google.com>

c-family/ChangeLog.pph
	* c-common.h (pph_disable_reader): Declare.
	* c-opts.c (pph_disable_reader): New.

cp/ChangeLog.pph
	* pph-core.c (pph_include_handler): If PPH has been disabled,
	process the file as a regular text include.
	* pph-out.c (pph_disable_output): Also disable the reader.

From-SVN: r183645