Add a GC root for the lexer.
This GC root is required to support running the garbage collector
during lexing.  This happens when the lexer loads pre-parsed headers.
Loading these headers calls into the middle-end, which may run the
garbage collector.

This was triggering in some of the bigger translation units that cause
a lot of GC activity.

2012-03-14   Diego Novillo  <dnovillo@google.com>

	* parser.c (the_lexer): Declare.
	(cp_lexer_new_main): Initialize.
	(c_parse_file): Set to NULL.

From-SVN: r185399