Fix x1mbstate_t.h.
This patch fixes the parser segmentation fault caused by a name
lookup failure (details in
http://gcc.gnu.org/ml/gcc-patches/2012-03/msg01369.html).

I am not 100% sure that this is the right fix, but Jason seems to
think that the theory behind this is fine (parser does no allow a
USING_DECL to be set in the bindings of an identifier).  I left a
FIXME note to help future debugging.

2012-03-21   Diego Novillo  <dnovillo@google.com>

cp/ChangeLog.pph
	* name-lookup.c (pph_set_namespace_decl_binding): Ignore
	USING_DECLs.

testsuite/ChangeLog.pph
	* g++.dg/pph/x1mbstate_t.h: Mark fixed.

From-SVN: r185607