Mangle TRAIT_EXPR and CONST_DECLs in template parameters.
We use the mangler to generate unique strings to represent types and symbols
that need to be merged.  However, the C++ mangler does not handle all
possible combinations.  This patch works around this problem by using
a combination of the pretty printer and the mangler.  This only works
when a PPH image is being generated, so it does not affect the
standard mangler.

2012-03-25   Diego Novillo  <dnovillo@google.com>

	* mangle.c (write_expression): Handle TRAIT_EXPR when PPH is enabled.
	(write_template_arg_literal): Handle CONST_DECLs that do not have
	a constant DECL_INITIAL when PPH is enabled.

From-SVN: r185808