Fix check for empty statement list.
We do some sanity checks before starting to write a PPH image.  The
check was inconsistent: for the VEC scope_chain->x_stmt_tree.x_cur_stmt_list
we were asserting that it was NULL *and* empty.

We just need it to be empty.  It doesn't matter if it was allocated.

2012-03-25   Diego Novillo  <dnovillo@google.com>

	* pph-out.c (pph_out_global_binding_keys): Fix check for empty
	statement list.

From-SVN: r185809