Added code to verify that verification for construction vtables is needed
Check if a  VTT is needed and only generate the verification for the refered
conctruction vtables if needed. There seems to be a bug in GCC where
some VTTs are being marked as needed so I also had to check for the
"finalized" flag. This needs to be reviewed.

From-SVN: r187121