Fixes a problem where we were still trying to output entries in the .debug_addr table for location lists that were removed...
Fixes a problem where we were still trying to output entries
in the .debug_addr table for location lists that were removed,
and fixes a problem where we were getting an ICE while trying
to output a pubname for a member function of a struct.

Tested with the GCC testsuite and validate-failures.py, and
also with an internal Google build with -gfission enabled.


2012-05-29   Sterling Augustine  <saugustine@google.com>
	     Cary Coutant  <ccoutant@google.com>

	* gcc/dwarf2out.c (remove_loc_list_addr_table_entries): New function.
	(is_class_die): Return TRUE for DW_TAG_structure_type.
	(resolve_addr): Remove address table entries when replacing a
	location list.

Co-Authored-By: Cary Coutant <ccoutant@google.com>

From-SVN: r187983