This patch works around a subtlety in the way weak references to symbols defined in archives are handled by the linker.
This patch works around a subtlety in the way weak references to
symbols defined in archives are handled by the linker. This is an
issue because google binaries are using weak references to detect
whether the libgcov.a version contains the new __gcov_reset and
__gcov_dump interfaces. Since they are defined in their own object
file within the archive, and no strong references are made to any
other symbols within the same object file, the linker does not
resolve the weak references. The workaround is to add dummy
strong references to the new routines from the main object file
(defined under L_gcov) that will always be referenced during
-fprofile-generate builds.

2012-06-01   Teresa Johnson  <tejohnson@google.com>

	* libgcov.c: Add references to gcov_reset and gcov_dump
        from L_gcov section.

From-SVN: r188134