This patch is for the google/gcc-4_6 branch.
In porting the earlier patch to use constant forms for DW_AT_high_pc,
the code in dwarf2out_finish was inadvertently changed to use an
indirect form for DW_AT_low_pc in the main compile_unit DIE.  This
patch adds the force_direct parameter used in other cases to
distinguish between DIEs in the .o (which should use the normal direct
form) and DIEs in the .dwo (which need to use DW_FORM_addr_index).

Tested with make check-gcc and validate-failures.py.


2012-06-07   Cary Coutant  <ccoutant@google.com>

	* gcc/dwarf2out.c (add_AT_low_high_pc): Add force_direct parameter;
	add DW_FORM_addr_index reference when splitting debug info.
	(gen_subprogram_die): Pass FALSE when calling add_AT_low_high_pc.
	(add_high_low_attributes): Likewise.
	(dwarf2out_finish): Pass TRUE when calling add_AT_low_high_pc.

From-SVN: r188313