This patch is for the google/main and google/gcc-4_7 branches.
New code in GCC 4.7 is calling linemap_lookup with a location_t that
may still represent a location-with-discriminator.  Before using a
location_t value to lookup the line number, it needs to be mapped to
a real location_t value.

Tested with make check-gcc and validate-failures.py.


2012-08-09   Cary Coutant  <ccoutant@google.com>

gcc/
	* tree-diagnostic.c (maybe_unwind_expanded_macro_loc): Check for
	discriminator.
	* diagnostic.c (diagnostic_report_current_module): Likewise.

From-SVN: r190269