This patch is for the google/gcc-4_7 branch.
This patch is for the google/gcc-4_7 branch.  I'll submit it for trunk
after the Fission patches have gone in.

When adding names to the pubnames table (-gsplit-dwarf or -gpubnames),
a method within a union may not get handled properly, sometimes resulting
in an internal compiler error in output_pubnames().

This patch fixes the problem by using the existing predicate,
class_scope_p, instead of is_class_die, which failed to test for
a union type.


2012-08-17   Cary Coutant  <ccoutant@google.com>

gcc/
	* dwarf2out.c (is_class_die): Remove function.
	(add_pubname): Call class_scope_p instead.

From-SVN: r190494